move nav diagram into folder
