+++ components/navdiagram/NavDiagram.tsx
