+++ components/navdiagram/NavDiagramCourseNode.tsx
