+++ components/navdiagram/NavDiagramSectionNode.tsx
