+++ components/navdiagram/NavDiagramThemeNode.tsx
